Rough partial prototype for pipelines.
No external commands yet in this version, which makes it not super
useful as is.  Also non-async, which means the first command runs to
completion with its output buffered, then the next command, etc.

But it's enough for exploring some thoughts I'd had last night
about the API.  Note the example:

    cat(b'/etc/shells') | split(lines=True)

(That's with a stub `cat` as a "shell function", having not implemented
external commands.)  I imagine this looking in shell syntax like

    sh { cat /etc/shells | split -l }

And the example works!  With pprint(list(...)):

    [b'# /etc/shells: valid login shells',
     b'/bin/sh',
     b'/bin/dash',
     b'/bin/bash',
     ...]