shwords: Add {!@} to expand a whole iterable.
Using a "conversion" `!@` rather than a "format spec" `:@` because the
former is where the Python `format` mechanism applies its own generic
logic, whereas the latter is delegated to the object to be formatted
via `__format__` (and its C API equivalent), and this feels like it
fits best as generic global logic.

Dropping the `r` and `a` conversions while here, because they seem
pretty Python-centric.  The `s` is maybe superfluous too, but it's at
least potentially useful for its original purpose: to take a Python
object of some random type that has a useful `__str__` but hasn't
grown a `__format__`, and convert it up front to a string.