shwords: Replace C parser with a few regexps.
The main point of this change is to give us better control of the loop,
so that we can do this:

    match = PAT_SPACE_OR_END.match(fmt, pos)
    # if not match: raise ...
    pos = match.end()

to directly advance the loop, in place of the rather convoluted
`after_list` mechanism.  We might find further improvements to make
using this greater control, too.

But also, this script gets 21 lines longer to replace about 200 lines
of C.  Not thoroughly tested so I'm sure there are bugs, but bugs aside,
the only functional change I know of is that error messages for invalid
format strings become less specific.  Pretty sure these regexps are just
a better way of describing this than the C code.