Fix a test so it works with default Git config. :-)
I have `git config --global core.abbrev 9`, because I prefer using
abbreviated commit IDs in all human contexts and want to keep it
clearly unnecessary to worry about them growing collisions.

That means a silly test like this needs to be explicit about that bit
of Git config, though, so that it gets the same results for me and in
a default configuration.