Add an example Bash script, and a translation to Python with pysh.
Excluding commentary added for this exercise, 47 lines of Bash script
became 55 lines of Python.  Not bad!

Several notes inline on things I noticed while doing the translation.

One factor in the script not growing too much is savings from letting
Click do the CLI-parsing: 14 lines of Bash became 8 lines of Python,
including `import click` and a couple of blank lines; so -6 lines net.

On the other side of the ledger, most of the expansion is in
boilerplate at start and end: imports, `if __name__` etc., and the
blank lines separating each of them from the stuff in between add up
to 13 lines.  Compared to `set -eu -o pipefail` plus a blank line,
that's +11 lines net expansion.  (Of which one line is double-counted
from above.)

I think the biggest other effect is handling of failure exits; these
come up in a few different ways, and all are verbosely handled by
`try`/`except` blocks catching `subprocess.CalledProcessError`.
That's +4 net lines in this example; would be more, but Bash itself
gets messy with `||` fallbacks inside a `$()` command substitution.