pysh.subprocess: Try that module-level docstring again.
In this version, I think the first paragraph alone, the first two, the
first three, etc., are all reasonably clear and self-contained without
yet reading the rest.

Naturally that's helpful if the information the reader wants happens
to be in the first few paragraphs.  But also I think it makes it much
easier to understand even when reading the whole thing, because while
you're partway through you don't have to hold in your head a bunch of
loose threads.

On top of that, it's a good bit shorter, while containing virtually
all the same information.