pysh.cmd: Rename `split(lines=True)` as `splitlines()`.
I haven't yet wanted to use `lines=False`.  And this implementation
didn't actually work correctly for that, because of the quite
different semantics of `bytes.split` (like `str.split`) when
splitting on whitespace vs. a given delimiter.

We can always add the feature later when there's a use case in
front of us, but no need to have it in the API until then.