pytest: Stop looking in all *.py files for tests.
We'll need to exclude setup.py when we add that shortly; and, as it
happens, I just recently (in be49f046f) moved all the tests that had
been embedded in implementation files out to specifically test files
anyway, so this is no longer needed.

If in the future we want tests in non-test files again, we can
always adjust this further.