pytest: Stop looking in all *.py files for tests.
We'll need to exclude setup.py; and, as it happens, I just recently
moved all the tests that had been embedded in implementation files
out to specifically test files anyway, so this is no longer needed.

If in the future we want tests in non-test files again, we can always
adjust this further.