Fix to work on Python 3.6
That was pretty easy!  Glad to have a nice test suite.

For the moment, testing consists of running a `pytest` that was
installed in a Python 3.6 virtualenv.  (I just run `pytest3.6`,
having set up such a virtualenv in ~/.local/venvs/ and symlinked
its `pytest` into ~/.local/bin/.)  Should automate that so it's
simple to routinely run on 3.6, 3.7, oh and also 3.8 and the
pre-release versions of 3.9.