doc: Absolutize internal links for the dist metadata
This way they actually work, when browsed on PyPI!  This wasn't an
issue before the previous commit, because these links were all in
the toplevel repo README, not the library README, and didn't get
included in the dist metadata.

Keeping them relative in the source means that when browsing on
GitHub, if you're at a specific commit (or a fork, etc.) they stay
at that commit.  Ideally we'd accomplish that for browsing on PyPI
too, but that seems trickier; just note it as a TODO.